Release/1.2

+++ backend/posts/views.py
@@ -10,8 +10,14 @@ from math import ceil
 from posts.models import Post, PostImage
 from users.models import User
 from tags.models import Tag, TagClass
+from groups.models import Group
+from workouts.models import Routine
+from django.db.models import Count
+
 
 from comments.views import prepare_comment_response
+from groups.views import prepare_compact_group_response
+from workouts.views import prepare_fitelements_list
 
 
 def add_exp(username, exp):
@@ -74,6 +80,16 @@ def prepare_post_response(post, is_detail, username):
             image_response.append(image["image"])
         response["images"] = image_response
 
+        response["group"] = prepare_compact_group_response(post.group)
+        routine_single = post.routine
+        if routine_single is not None:
+            response["routine"] = {
+                "id": routine_single.id,
+                "author": routine_single.author.id,  # id or name
+                "name": routine_single.name,
+                "fitelements": prepare_fitelements_list(routine_single.fit_element.all()),
+            }
+
     return response
 
 
@@ -100,7 +116,7 @@ def post_home(request):
         offset = (query_args["page_num"] - 1) * query_args["page_size"]
         limit = query_args["page_num"] * query_args["page_size"]
 
-        posts = Post.objects.all()
+        posts = Post.objects.filter(in_group=None)
 
         if query_args["keyword"]:
             filter_args = {}
@@ -137,22 +153,55 @@ def post_home(request):
                 if ("prime_tag" in data.keys() and data["prime_tag"])
                 else None
             )
+            in_group = (
+                Group.objects.get(pk=data["group_id"]) if ("group_id" in data.keys()) else None
+            )
+            routine = Routine.objects.get(pk=data["routine"]) if (data["routine"] != '') else None
+            group = Group.objects.get(pk=data["group"]) if (data["group"] != '') else None
 
             created_post = Post.objects.create(
-                author=author, title=data["title"], content=data["content"], prime_tag=prime_tag
+                author=author,
+                title=data["title"],
+                content=data["content"],
+                prime_tag=prime_tag,
+                in_group=in_group,
+                routine=routine,
+                group=group,
             )
+
             for tag in data["tags"]:
                 tag = Tag.objects.get(pk=tag["id"])
                 created_post.tags.add(tag)
             for image in data["images"]:  # image would be string type
                 PostImage.objects.create(image=image, post=created_post)
 
-            add_exp(request.user.username, 20)
+            add_exp(request.user.username, 10)
 
             return JsonResponse({"post_id": str(created_post.pk)}, status=201)
             # data should have user, post info.
         except (KeyError, json.JSONDecodeError, User.DoesNotExist, Tag.DoesNotExist):
             return HttpResponseBadRequest()
+
+
+@require_http_methods(["GET"])
+def post_group(request, group_id):
+    try:
+        group = Group.objects.get(pk=group_id)
+        group.members.get(username=request.user.username)
+    except (Group.DoesNotExist, User.DoesNotExist):
+        return HttpResponseNotFound()
+
+    posts = Post.objects.filter(in_group=group)
+    posts_serial = prepare_posts_response(posts)
+
+    # Total page number calculation.
+    response = JsonResponse(
+        {
+            "posts": posts_serial,
+        },
+        status=200,
+    )
+    return response
 
 
 @require_http_methods(["GET", "PUT", "DELETE"])
@@ -166,6 +215,13 @@ def post_detail(request, query_id):
         try:
             post_id = int(query_id)
             post_obj = Post.objects.get(pk=post_id)
+
+            if post_obj.in_group:
+                try:
+                    post_obj.in_group.members.get(username=request.user.username)
+                except User.DoesNotExist:
+                    return HttpResponseNotFound()
+
             return JsonResponse(
                 prepare_post_response(post_obj, True, request.user.username), status=200
             )
@@ -176,6 +232,9 @@ def post_detail(request, query_id):
             data = json.loads(request.body.decode())
             post_id = int(query_id)
             post_obj = Post.objects.get(pk=post_id)
+
+            if post_obj.author.username != request.user.username:
+                return HttpResponseBadRequest()
 
             post_obj.title = data["title"]
             post_obj.content = data["content"]
@@ -202,6 +261,12 @@ def post_detail(request, query_id):
                 if image.image not in data["images"]:
                     image.delete()
 
+            # Routine & Group
+            post_obj.routine = (
+                Routine.objects.get(pk=data["routine"]) if (data["routine"] != '') else None
+            )
+            post_obj.group = Group.objects.get(pk=data["group"]) if (data["group"] != '') else None
+
             post_obj.save()
             return JsonResponse({"message": "success"}, status=200)
         except Post.DoesNotExist:
@@ -212,6 +277,9 @@ def post_detail(request, query_id):
         try:
             post_id = int(query_id)
             post_obj = Post.objects.get(pk=post_id)
+
+            if post_obj.author.username != request.user.username:
+                return HttpResponseBadRequest()
 
             post_obj.delete()
             return JsonResponse({"message": "success"}, status=200)
@@ -288,3 +356,14 @@ def post_func(request, query_id):
         return JsonResponse({"type": type_of_work}, status=200)
     except (Post.DoesNotExist, User.DoesNotExist):
         return HttpResponseNotFound()
+
+
+@require_http_methods(["GET"])
+def post_main(request):
+    """
+    GET : get data for main page.
+    """
+    posts = prepare_posts_response(
+        Post.objects.annotate(num_likes=Count('liker')).all().order_by('-num_likes')[:5]
+    )
+    return JsonResponse({"posts": posts}, status=200)
